// src/utils/types.ts
export type ColorQuantity = { color: string; quantity: number }
export type Variant = { size: string; colors: ColorQuantity[] }
// ...
export type CartItemType = {
  images: string[]
  name: string
  price: number
  color: string
  size: string
  availableVariants: Variant[]
  amount: number
  id: string
  vendor: string
  vendorid: string
  category: string
}